requirements.txt
open_clip
torch
PIL
tqdm
numpy
torchvision
scikit-learn
opencv-python
pyyaml
scipy
imgaug
